pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["sfkan6 <sfkan6@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
django-mptt = "^0.13.4"
Pillow = "^9.2.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^6.3.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
